Rename FindCMinpack.cmake to CMinpackConfig.cmake to use the 'config' mode when calling find_package()
The documentation of find_package() states that it operates in 2 modes:

- Module
- Config

The 'Module' mode should be used for modules that come bundled with cmake.

The 'Config' mode should be used for modules that are installed in the system.
This is evident when we see that some installations of cmake bundle the version
in the directory name (eg.: cmake-3.5).

The current state of the code fails for such installations, since when
find_package(CMinpack) is called, cmake will fail to find the FindCMinpack.cmake
file since it is installed in 'share/cmake/Modules', but cmake is searching
'share/cmake-3.5/Modules'.

If the file is installed in 'share/cmake/CMinpack' (as proposed in the commit), all
cmake versions will be able to find the file in 'Config' mode.

The documentation of find_package has more info: https://cmake.org/cmake/help/v3.5/command/find_package.html