Remove leftover noise from VERSION/SOVERSION variables
Commit 56088f9bf6c2 removed the CACHE keyword from the
set command, which made the trailing help text part of the
variable values. This extra text was propagated to the
pkgconfig file, making it malformed.

As there is no use for any help text (e.g. in CMake GUI) for a
variable which can not be overridden by a user, remove the
trailing text.

Fixes #55